+++ 2019/src/day9.h
@@ -31,7 +31,7 @@ public:
     std::ifstream inputFile{input_path};
 
     for (std::string line; std::getline(inputFile, line, ',');) {
-      program.push_back(std::stoi(line));
+      program.push_back(std::stol(line));
     }
     return program;
   }
